--- nextstep-mvc/src/main/java/nextstep/mvc/ControllerHandlerMapping.java
@@ -1,7 +1,0 @@
-package nextstep.mvc;
-
-import nextstep.mvc.asis.Controller;
-
-public interface ControllerHandlerMapping extends HandlerMapping {
-    Controller getHandler(String requestUri);
-}
