+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,19 +1,30 @@
 package nextstep.mvc;
 
-import nextstep.mvc.tobe.exception.RequestUrlNotFoundException;
-
+import java.io.IOException;
+import java.util.List;
+import java.util.Objects;
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import nextstep.mvc.asis.Controller;
+import nextstep.mvc.tobe.HandlerExecution;
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.exception.RequestUrlNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 @WebServlet(name = "dispatcher", urlPatterns = "/", loadOnStartup = 1)
 public class DispatcherServlet extends HttpServlet {
+    private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
+    private static final String DEFAULT_REDIRECT_PREFIX = "redirect:";
     private static final long serialVersionUID = 1L;
-    private HandlerMapping[] handlerMappings;
+    private List<HandlerMapping> handlerMappings;
 
-    public DispatcherServlet(HandlerMapping... handlerMappings) {
+    public DispatcherServlet(List<HandlerMapping> handlerMappings) {
         this.handlerMappings = handlerMappings;
     }
 
@@ -26,11 +37,36 @@ public class DispatcherServlet extends HttpServlet {
 
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException {
-        for (HandlerMapping handlerMapping : handlerMappings) {
-            if (handlerMapping.handle(req, resp)) {
+        try {
+            Controller controller = (Controller) handlerMappings.get(0).getHandler(req);
+            if (Objects.nonNull(controller)) {
+                String viewName = controller.execute(req, resp);
+                move(viewName, req, resp);
                 return;
             }
+
+            HandlerExecution execution = (HandlerExecution) handlerMappings.get(1).getHandler(req);
+            if (Objects.nonNull(execution)) {
+                ModelAndView modelAndView = execution.handle(req, resp);
+                modelAndView.render(req, resp);
+                return;
+            }
+        } catch (Exception e) {
+            logger.error("Exception : {}", e);
+            throw new ServletException(e.getMessage());
         }
+
         throw new RequestUrlNotFoundException();
+    }
+
+    private void move(String viewName, HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        if (viewName.startsWith(DEFAULT_REDIRECT_PREFIX)) {
+            resp.sendRedirect(viewName.substring(DEFAULT_REDIRECT_PREFIX.length()));
+            return;
+        }
+
+        RequestDispatcher rd = req.getRequestDispatcher(viewName);
+        rd.forward(req, resp);
     }
 }
