+++ nextstep-mvc/src/main/java/nextstep/mvc/HandlerMapping.java
@@ -1,11 +1,9 @@
 package nextstep.mvc;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 public interface HandlerMapping {
     void initialize();
 
-    boolean handle(HttpServletRequest req, HttpServletResponse resp) throws ServletException;
+    Object getHandler(HttpServletRequest request);
 }
