--- nextstep-mvc/src/main/java/nextstep/mvc/ModelAndViewHandlerMapping.java
@@ -1,9 +1,0 @@
-package nextstep.mvc;
-
-import nextstep.mvc.tobe.HandlerExecution;
-
-import javax.servlet.http.HttpServletRequest;
-
-public interface ModelAndViewHandlerMapping extends HandlerMapping {
-    HandlerExecution getHandler(HttpServletRequest req);
-}
