+++ slipp/src/main/java/slipp/ManualHandlerMapping.java
@@ -1,24 +1,18 @@
 package slipp;
 
-import nextstep.mvc.ControllerHandlerMapping;
+import java.util.HashMap;
+import java.util.Map;
+import javax.servlet.http.HttpServletRequest;
+
+import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.asis.Controller;
 import nextstep.mvc.asis.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import slipp.controller.*;
 
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-
-public class ManualHandlerMapping implements ControllerHandlerMapping {
+public class ManualHandlerMapping implements HandlerMapping {
     private static final Logger logger = LoggerFactory.getLogger(ManualHandlerMapping.class);
-    private static final String DEFAULT_REDIRECT_PREFIX = "redirect:";
     private Map<String, Controller> mappings = new HashMap<>();
 
     @Override
@@ -38,43 +32,8 @@ public class ManualHandlerMapping implements ControllerHandlerMapping {
     }
 
     @Override
-    public boolean handle(final HttpServletRequest req, final HttpServletResponse resp) throws ServletException {
-        String requestUri = req.getRequestURI();
-        logger.debug("Method : {}, Request URI : {}", req.getMethod(), requestUri);
-
-        Controller controller = getHandler(requestUri);
-        if (doesNotExistsController(controller)) {
-            return false;
-        }
-
-        try {
-            String viewName = controller.execute(req, resp);
-            move(viewName, req, resp);
-        } catch (Throwable e) {
-            logger.error("Exception : {}", e);
-            throw new ServletException(e.getMessage());
-        }
-        return true;
-    }
-
-    private boolean doesNotExistsController(final Controller controller) {
-        return Objects.isNull(controller);
-    }
-
-    private void move(String viewName, HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        if (viewName.startsWith(DEFAULT_REDIRECT_PREFIX)) {
-            resp.sendRedirect(viewName.substring(DEFAULT_REDIRECT_PREFIX.length()));
-            return;
-        }
-
-        RequestDispatcher rd = req.getRequestDispatcher(viewName);
-        rd.forward(req, resp);
-    }
-
-    @Override
-    public Controller getHandler(String requestUri) {
-        return mappings.get(requestUri);
+    public Object getHandler(HttpServletRequest req) {
+        return mappings.get(req.getRequestURI());
     }
 
     void put(String url, Controller controller) {
